Introduce builtin function go and makechan
Builtin go(fn, arg1, arg2, ...) starts a goroutine which run
fn(arg1, arg2, ...) in a new VM cloned from the current running VM,
and returns a job object that has wait, result, abort methods.

Builtin makechan(size) accepts an optional size parameter, makes a
channel to send/receive object, and returns a chan object that has
send, recv, close methods.

To be able to access *VM instance that it is running in for go()
and makechan(), now VM will always pass VMObj to and only to
BuiltinFunction as arg[0].