govm now supports calling both script functions and native functions
with this commit:

1. builtinGovm() now accepts native golang function, but unlike
compiled functions which run in a cloned VM, native golang functions
do not need VM, they just run in a goroutine.

2. Parent VMs now have records of errors of all its descendent VMs.
A VM's error is a combination of the error of itself and all errors
of the descendent VMs, in another word, parent VM's error includes
all errors of the child VMs.

3. ErrVMAborted is no longer treated as runtime error.