improve panic handling
On a panic happens in a VM:
* the panic value along with both its script level and native level callstack
are recorded as run time error, its parent VM should examine this error.
* the panic will not cause whole program exits, but it does trigger the abort
chain, so that the VM in which the panic happens and all its descendent VMs
will terminate.

So panic only causes the descendent group of VMs to exit.
Other VM groups are not affected.